Fix PostgreSQL compatibility issues and improve Render deployment
- Batch card inserts (200 per query) for ~100x faster sync with remote Neon DB
- Fix INTEGER type errors: round decimal values for pop_rank, prices columns
- Fix meta_snapshots.pop_rank schema: INTEGER → REAL to accept decimal values
- Fix ecosystem query: cast snapshot_date to date for timestamp comparison
- Fix GROUP BY LOWER(name) PostgreSQL syntax in duplicate cleanup
- Add "type": "module" to package.json for ESM support
- Copy schema.sql to dist/ during build for production deployment

+++ server/src/db/schema.sql
@@ -87,7 +87,7 @@ CREATE TABLE IF NOT EXISTS meta_snapshots (
   deck_type_name TEXT NOT NULL,
   tier INTEGER,
   power REAL,
-  pop_rank INTEGER,
+  pop_rank REAL,
   snapshot_date TEXT NOT NULL,
   UNIQUE(deck_type_name, snapshot_date)
 );
